% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aicplot.R
\name{aicplot}
\alias{aicplot}
\title{Plot AIC values}
\usage{
aicplot(cpobj, dp.aic = 2)
}
\arguments{
\item{cpobj}{list, contains a vector of AIC values}

\item{dp.aic}{numeric, digits for the AIC values}
}
\value{
returns a plot which shows the AIC values of the estimating
procedure
}
\description{
Create plot of AIC values of estimating procedure
}
\examples{
# Create a vector with AIC values
AIC_values <- c(1950:1910, 1910:1920, 1920:1880, 1880:1920)
AIC_values <- round(AIC_values + rnorm(length(AIC_values),
                    mean = 0, sd = 5), digits = 2)

cpobj <- list(AIC_values = AIC_values)
aicplot(cpobj, dp.aic = 2)
}
\seealso{
\code{\link{est.cutpoint}}
}
