% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.factors.R
\name{combine.factors}
\alias{combine.factors}
\title{Combine Factors}
\usage{
combine.factors(bandwith = 0.1, nbofcp = 1, nrm, symtails = FALSE)
}
\arguments{
\item{bandwith}{numeric, determines the minimum size per group of the
dichitomised biomarker}

\item{nbofcp}{numeric, number of cutpoints searching for}

\item{nrm}{numeric, number of rows in cpdata after removing observations
with missing values in biomarker}

\item{symtails}{logical, if TRUE the tails of the dichotomised biomarker
are symmetrical}
}
\value{
returns an integer matrix with all factor combinations of the
dichotomised biomarker
}
\description{
Create matrix with all factor combinations of the biomarker
}
