% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_factors.R
\name{combine_factors}
\alias{combine_factors}
\title{Combine Factors}
\usage{
combine_factors(bandwith = 0.1, nb_of_cp = 1, nrm, symtails = FALSE)
}
\arguments{
\item{bandwith}{numeric, determines the minimum size per group of the
dichitomised biomarker}

\item{nb_of_cp}{numeric, number of cutpoints searching for}

\item{nrm}{numeric, number of rows in cpdata after removing observations
with missing values in biomarker}

\item{symtails}{logical, if TRUE the tails of the dichotomised biomarker
are symmetrical}
}
\value{
Returns all factor combinations of the dichotomized variable.
Create matrix with all factor combinations of the dichotomised biomarker
Delete rows with less than ‘bandwith’ percent of the number of columnsconvert percentages into number

If nb_of_cp == 1
How often does the 1 occur per line
Delete rows with less than ‘bandwith’ percent of the number of columns
If nb_of_cp == 2
Delete all rows with less than bandwith per cent of the nb. of columns
How often does the 2 occur per line
If m.perm.fun should be symmetrical
}
\description{
Create matrix with all factor combinations of the biomarker
}
