% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data1}
\alias{data1}
\title{Dataset for testing the cutpoint estimating function: est_cutpoint}
\format{
\subsection{"data1"}{

A data frame with 100 rows and 5 variables:
\describe{
\item{biomarker}{numeric from 1 to 257}
\item{covariate_1}{numeric, from 4.25 to 12.33, with effect of the cutpoint
of the biomarker}
\item{covariate_2}{numeric, from 465 to 1205, with no or small effect of
the cutpoint of the biomarker}
\item{time}{numeric, from 3 to 328}
\item{event}{numeric, 0 or 1}
}
}
}
\source{
Self-generated example data
}
\usage{
data(data1)
}
\description{
A dataset containing data for testing the estimating of one or two cutpoints
}
\examples{
data(data1)

}
\author{
Jan Porthun
}
\keyword{datasets}
