% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data2_ushape}
\alias{data2_ushape}
\title{Dataset for testing the ushape-cutpoint function}
\format{
\subsection{"data2_ushape"}{

A data frame with 200 rows and 4 variables:
\describe{
\item{biomarker}{numeric from 1e-04 to 4.7}
\item{covariate_1}{numeric, from 8.07e-05 to 1.90}
\item{time}{numeric, from 0.002 to 5.09}
\item{event}{numeric, 0 or 1}
}
}
}
\source{
Self-generated example data
}
\usage{
data(data2_ushape)
}
\description{
A dataset containing data for testing the ushape-cutpoint function.
}
\examples{
data(data2_ushape)

}
\author{
Jan Porthun
}
\keyword{datasets}
