% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pspline.plot.R
\name{pspline.plot}
\alias{pspline.plot}
\title{Plot psplines}
\usage{
pspline.plot(cpobj, show.splines = TRUE)
}
\arguments{
\item{cpobj}{list contains variables for pspline plot}

\item{show.splines}{logical, if TRUE shows the splines with different DF}
}
\value{
returns a plot with the pspline and shows the cutpoints
}
\description{
Create pspline plot with different degrees of freedom and
shows the cutpoints of the biomarker
}
\examples{
biomarker <- rnorm(100, mean = 100, sd = 10)
time <- seq(1, 100, 1)
event <- rbinom(100, 1, 0.5)
datf <- data.frame(time, event, biomarker)
plot_splines_list <- list(cpdata = datf, nbofcp = 1, cp = 95, dp = 2,
    cpvarname = "Biomarker")
pspline.plot(plot_splines_list)
}
\references{
Harrell, F. E. (2001). Regression modeling strategies: with applications to linear models, logistic and ordinal regression, and survival analysis. Springer Science & Business Media.

Get quantiles of biomarker
Get optimal degree of freedom
Define degree of freedom used for termplot
}
\seealso{
\code{\link{est.cutpoint}}
}
\keyword{cutpoint}
\keyword{plot}
\keyword{pspline}
\keyword{termplot}
\keyword{visualization}
