% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pspplot.R
\name{pspplot}
\alias{pspplot}
\title{Plot psplines}
\usage{
pspplot(cpobj, show.splines = TRUE)
}
\arguments{
\item{cpobj}{list contains variables for pspline plot}

\item{show.splines}{logical, if TRUE shows the splines with different DF}
}
\value{
returns a plot with the pspline and shows the cutpoints
}
\description{
Create pspline plot with different degrees of freedom and
shows the cutpoints of the biomarker
}
\examples{
biomarker <- rnorm(100, mean = 100, sd = 10)
time <- seq(1, 100, 1)
event <- rbinom(100, 1, 0.5)
datf <- data.frame(time, event, biomarker)
plot_splines_list <- list(cpdata = datf, nbofcp = 1, cp = 95, dp = 2,
    cpvarname = "Biomarker")
pspplot(plot_splines_list)
}
\keyword{Define}
\keyword{Get}
\keyword{biomarker}
\keyword{cutpoint}
\keyword{degree}
\keyword{for}
\keyword{freedom}
\keyword{of}
\keyword{optimal}
\keyword{plot}
\keyword{pspline}
\keyword{quantiles}
\keyword{termplot}
\keyword{used}
\keyword{visualization}
