% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_plot.R
\name{value_plot}
\alias{value_plot}
\title{Plot AIC and LRT values}
\usage{
value_plot(
  cpobj,
  plotvalues = "AIC",
  dp.plot = 2,
  show_limit = TRUE,
  plottype2cp = "contour"
)
}
\arguments{
\item{cpobj}{list, contains a vector of AIC values (AIC_values) and
Likelihood ratio test values (LRT_values) of the estimating procedure}

\item{plotvalues}{character, either "AIC" or "LRT". Either the AIC or
LRT values are displayed. Default is "AIC".}

\item{dp.plot}{numeric, digits for the AIC values and LRT values.
Default is 2.}

\item{show_limit}{logical, if TRUE the minimum AIC value is shown in the plot
if plotvalues = "AIC" and the maximum LRT value is shown if
plotvalues = "LRT"}

\item{plottype2cp}{character, either "contour" or "index". Default is
"contour".This option is available only when searching for two cutpoints.
Index plots and contour plots can be selected. Index plots display all AIC
or LRT values from the estimation process as a scatter plot. Contour plots
are shown in the RStudio viewer and illustrate the two potential cutpoints
along with the corresponding AIC or LRT values. Index plots that do not
show extreme values suggest that there may not be any actual cutpoints in
the data. Contour plots provide an opportunity to explore whether there
could be other potential cutpoints with similar AIC or LRT values.}
}
\value{
Plots the AIC- or LRT-values, derived from the estimation procedure.
}
\description{
Create plot of AIC- or Likelihood ratio test values of the
estimating procedure
}
\examples{
# Print the AIC-values
# Create AIC values:
AIC_values <- c(1950:1910, 1910:1920, 1920:1880, 1880:1920)
AIC_values <- round(AIC_values + rnorm(length(AIC_values),
                    mean = 0, sd = 5), digits = 2)

# Create a cutpoint variable:
cpvariable_values <- matrix(NA, nrow = length(AIC_values), ncol = 2)
cpvariable_values[ ,1] <- c(1:length(AIC_values))

# Create a cutpoint object (cpobj):
cpobj <- list(AIC_values        = AIC_values,
              nb_of_cp          = 1,
              cpvariable_values = cpvariable_values,
              cpvarname         = "Cutpoint variable"
              )

value_plot(cpobj, plotvalues = "AIC", dp.plot = 2, show_limit = TRUE)
}
\seealso{
\code{\link{est_cutpoint}}
}
