% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutpoint-package.R
\docType{package}
\name{cutpoint-package}
\alias{cutpoint}
\alias{cutpoint-package}
\title{cutpoint: Estimate cutpoints of metric variables in the context of Cox-regression}
\description{
One or two cutpoints of a metric variable, such as a biomarker, can be determined in the context of survival analysis. These cutpoints are utilized to create groups with differend survival probabilities.
}
\author{
\strong{Maintainer}: Jan Porthun \email{jan.porthun@ntnu.no}

}
\keyword{internal}
