% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutpoint-package.R
\docType{package}
\name{cutpoint-package}
\alias{cutpoint}
\alias{cutpoint-package}
\title{cutpoint: Estimate cutpoints of metric variables in the context of Cox-regression}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Estimate one or two cutpoints of a metric or ordinal-scaled variable in the multivariable context of survival data or time-to-event data. Visualise the cutpoint estimation process using contour plots, index plots, and spline plots. It is also possible to estimate cutpoints based on the assumption of a U-shaped or inverted U-shaped relationship between the predictor and the hazard ratio.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jan-por/cutpoint}
  \item Report bugs at \url{https://github.com/jan-por/cutpoint/issues}
}

}
\author{
\strong{Maintainer}: Jan Porthun \email{jan.porthun@ntnu.no} (\href{https://orcid.org/0000-0002-4100-4361}{ORCID})

}
\keyword{internal}
