% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est.cutpoint.R
\name{est.cutpoint}
\alias{est.cutpoint}
\title{Estimates cutpoints}
\usage{
est.cutpoint(
  cpvarname,
  time = "time",
  event = "event",
  cofactors,
  data = data,
  nbofcp = 1,
  bandwith = 0.1,
  ushape = FALSE,
  symtails = FALSE,
  dp = 2,
  plot_splines = TRUE,
  all_splines = TRUE
)
}
\arguments{
\item{cpvarname}{character, name of the variable for which the cutpoints are
estimated}

\item{time}{character, name of the time variable}

\item{event}{character, name of the event variable}

\item{cofactors}{character, vector with the names of the covariates or/ and
factors}

\item{data}{data.frame, data set with the variables}

\item{nbofcp}{numeric, number of cutpoints searching for}

\item{bandwith}{numeric, minimum group size in percent of the total sample
size, bandwith must be between 0 and 0.3}

\item{ushape}{logical, if TRUE the cutpoints are estimated with three
categories}

\item{symtails}{logical, if TRUE the cutpoints are estimated with symmetric
tails}

\item{dp}{numeric, number of decimal places the cutpoints are rounded to}

\item{plot_splines}{logical, if TRUE creates pspline plot}

\item{all_splines}{logical, if TRUE all splines are shown}
}
\value{
returns the estimated cutpoints and the characteristics of the groups
}
\description{
One or two cutpoints of a variable are estimated by the AIC
criterion in a Cox proportional hazards model. The cutpoints are estimated
by dichotomizing the variable and testing the log-likelihood ratio test.
The cutpoints with the lowest AIC value are chosen.
}
