% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors_combine.R
\name{factors_combine}
\alias{factors_combine}
\title{Combine Factors}
\usage{
factors_combine(bandwith = 0.1, nb_of_cp = 1, nrm, symtails = FALSE)
}
\arguments{
\item{bandwith}{numeric, determines the minimum size per group of the
dichitomised variable}

\item{nb_of_cp}{numeric, number of cutpoints searching for}

\item{nrm}{numeric, number of rows in cpdata after removing observations
with missing values in biomarker}

\item{symtails}{logical, if TRUE the tails of the dichotomised variable
are symmetrical}
}
\value{
All factor combinations of the dichotomized variable.
}
\description{
Intern function, used for creation of a matrix with all factor
combinations of the cutpoint-variable
}
\keyword{internal}
